+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added field `fivetran_connector.auth.key_id`.
   - Added field `fivetran_connector.auth.team_id`.
   - Added field `fivetran_connector.auth.client_secret`.
+- Resource `fivetran_destination` updates:
+    - Added field `fivetran_destination.config.workspace_name` for OneLake.
+    - Added field `fivetran_destination.config.lakehouse_name` for OneLake.
 
 ## [1.1.3](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.2...1.1.3)